feat(splash): Added an initial draft/WIP for the splashscreen
Since the 3D models are quite heavy (especially w/ slow 3G)
is not really practical to not show any content to the user.
For these reason I implemented a simple Splashscreen with a
spinner that will render something on the user screen while
the remaining content gets streamed basck using Suspense.