fix(all): Nevermind it seems that I have fixed everything
Indeed it seems like that adding 'use client' everywhere does is, in
fact, the correct approach when doing a migration like this one.

Next up I'm gonna start moving text logic in server components so that
they can be generated statically ahed of time and returned immediately
while the 3D model are fetched and wahtever.